move TextBounds out of text2d
